Pre-close with nested confirm
Update to allow for a nested confirm within a confirm.  Previously was
using window.confirm to block execution flow, but now uses the returned
promise to determine whether to keep the parent dialog open based on the
nested confirm's resolve/reject.  Also changed the pre-close to not be
called when calling a confirm dialog's confirm method.