Broadcast event when adding padding
When adding or removing a dialog, ngDialog will calculate the width of
the sidebar and add/remove this width as padding to the right side of
the body tag so the page doesn't jog to the side when overflow: hidden
is applied.  Now an event is broadcast whenever this padding is added or
removed with the pixel width as an argument so applications can choose
to add this padding to other elements that may need it (specifically
position: fixed elements).